pyproject.toml
[project]
name = "zillow-estimate-agent"
version = "0.1.0"
description = "Production-grade Python agent to fetch current Zillow Zestimate by address."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "fastapi>=0.116.1",
  "playwright>=1.52.0",
  "pydantic>=2.11.3",
  "python-dotenv>=1.0.1",
  "typer>=0.15.2",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
]
apify = [
  "apify-client>=1.8.1",
]

[project.scripts]
zestimate-agent = "zestimate_agent.cli:app"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
